Apply patch from Mats Kindahl for fixing pg upgrade
From 63d116f208c770fe388e6d9e643d66957c8e4b25 Mon Sep 17 00:00:00 2001
From: Mats Kindahl <mats@timescale.com>
Date: Mon, 4 May 2020 12:46:01 +0200
Subject: [PATCH] Don't rely on timescaledb.restoring for upgrade

If a binary upgrade is in progress (when using `pg_upgrade`) the
per-database setting of `timescaledb.restoring` can be included in the
dump, which causes `pg_upgrade` to fail.

This commit fixes this by checking the global variable
`IsBinaryUpgrade` and not refreshing cache if we are in the middle of
doing a binary upgrade.